re-organize cli tools and modify option name
