+++ .gitignore
@@ -1,3 +1,5 @@
+const/
+model/
 *nogit*
 *.DS_Store
 *sandbox*
