--- tests/epitope.tsv
@@ -1,5 +1,0 @@
-PDFSSGGAAAAAAP
-AVATTQANT
-CCDIYKLV
-TRETDIAQGVP
-NFIELSFLVLKCF
